Basic router config
The basic config files of the 4 routers used for the lab. This configuration enable connectivity between all the networks. It also allow ssh and/or telnet connectivity from one router to all the others. The 2600 routers don't permit ssh connexion that is why telnet is configured.
The configuration is based on a real hardward network. The port connections are imposed by the type and number of ports available on the routers.